requirements.txt
mcp-server
gradio
pandas
numpy
scapy
python-nmap
